streamingccl: rename streamproducer to producer
Release note: none.
Epic: none.
