+++ pkg/BUILD.bazel
@@ -99,7 +99,7 @@ ALL_TESTS = [
     "//pkg/ccl/streamingccl/replicationutils:replicationutils_test",
     "//pkg/ccl/streamingccl/streamclient:streamclient_test",
     "//pkg/ccl/streamingccl/physical:physical_test",
-    "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
+    "//pkg/ccl/streamingccl/producer:producer_test",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
     "//pkg/ccl/testccl/authccl:authccl_test",
     "//pkg/ccl/testccl/sqlccl:sqlccl_test",
@@ -969,8 +969,8 @@ GO_TARGETS = [
     "//pkg/ccl/streamingccl/streamclient:streamclient_test",
     "//pkg/ccl/streamingccl/physical:physical",
     "//pkg/ccl/streamingccl/physical:physical_test",
-    "//pkg/ccl/streamingccl/streamproducer:streamproducer",
-    "//pkg/ccl/streamingccl/streamproducer:streamproducer_test",
+    "//pkg/ccl/streamingccl/producer:producer",
+    "//pkg/ccl/streamingccl/producer:producer_test",
     "//pkg/ccl/streamingccl:streamingccl",
     "//pkg/ccl/telemetryccl:telemetryccl_test",
     "//pkg/ccl/testccl/authccl:authccl_test",
