+++ pkg/ccl/ccl_init.go
@@ -33,7 +33,7 @@ import (
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/storageccl/engineccl"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/logical"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/physical"
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamproducer"
+	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/producer"
 	"github.com/cockroachdb/cockroach/pkg/ccl/utilccl"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/workloadccl"
 	"github.com/cockroachdb/cockroach/pkg/server"
