+++ pkg/ccl/streamingccl/physical/BUILD.bazel
@@ -26,7 +26,7 @@ go_library(
         "//pkg/ccl/streamingccl",
         "//pkg/ccl/streamingccl/replicationutils",
         "//pkg/ccl/streamingccl/streamclient",
-        "//pkg/ccl/streamingccl/streamproducer",
+        "//pkg/ccl/streamingccl/producer",
         "//pkg/ccl/utilccl",
         "//pkg/cloud/externalconn",
         "//pkg/cloud/externalconn/connectionpb",
@@ -127,7 +127,7 @@ go_test(
         "//pkg/ccl/streamingccl/replicationtestutils",
         "//pkg/ccl/streamingccl/replicationutils",
         "//pkg/ccl/streamingccl/streamclient",
-        "//pkg/ccl/streamingccl/streamproducer",
+        "//pkg/ccl/streamingccl/producer",
         "//pkg/cloud/impl:cloudimpl",
         "//pkg/cloud/nodelocal",
         "//pkg/jobs",
