+++ pkg/ccl/streamingccl/physical/stream_ingestion_job_test.go
@@ -19,7 +19,7 @@ import (
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl"
 	"github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/replicationtestutils"
 	"github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/replicationutils"
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamproducer"
+	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/producer"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/keys"
