+++ pkg/ccl/streamingccl/producer/event_stream.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package streamproducer
+package producer
 
 import (
 	"context"
