+++ pkg/ccl/streamingccl/producer/replication_manager.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package streamproducer
+package producer
 
 import (
 	"context"
