+++ pkg/ccl/streamingccl/producer/replication_stream_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package streamproducer_test
+package producer_test
 
 import (
 	"bytes"
