+++ pkg/testutils/lint/lint_test.go
@@ -2471,8 +2471,8 @@ func TestLint(t *testing.T) {
 			":!ccl/streamingccl/physical/replication_random_client_test.go",
 			":!ccl/streamingccl/physical/stream_ingestion_job_test.go",
 			":!ccl/streamingccl/physical/stream_ingestion_processor_test.go",
-			":!ccl/streamingccl/streamproducer/producer_job_test.go",
-			":!ccl/streamingccl/streamproducer/replication_stream_test.go",
+			":!ccl/streamingccl/producer/producer_job_test.go",
+			":!ccl/streamingccl/producer/replication_stream_test.go",
 			":!ccl/workloadccl/allccl/all_test.go",
 			":!cli/democluster/demo_cluster.go",
 			":!cli/democluster/demo_cluster_test.go",
